Adding modules to frontend code
