+++ GapsWeb/src/main/resources/templates/libraries.html
@@ -152,6 +152,6 @@
 <script src="/js/bootstrap.bundle.min.js" type="text/javascript"></script>
 <script src="/js/datatables.min.js" type="text/javascript"></script>
 <script src="/js/handlebars.js" type="text/javascript"></script>
-<script src="/js/libraries.js" type="text/javascript"></script>
+<script type="module" src="/js/libraries.js"></script>
 </body>
 </html>
