+++ GapsWeb/src/main/resources/templates/recommended.html
@@ -111,7 +111,8 @@
             <div class="progress top-margin" id="progressContainer">
                 <div aria-valuemax="100" aria-valuemin="0"
                      aria-valuenow="75"
-                     class="progress-bar progress-bar-striped progress-bar-animated" role="progressbar" style="width: 100%"></div>
+                     class="progress-bar progress-bar-striped progress-bar-animated" role="progressbar"
+                     style="width: 100%"></div>
             </div>
         </div>
 
@@ -170,6 +171,6 @@
 <script src="/js/sockjs-1.4.0.min.js" type="text/javascript"></script>
 <script src="/js/stomp-2.3.3.min.js" type="text/javascript"></script>
 <script src="/js/handlebars.js" type="text/javascript"></script>
-<script src="/js/recommended.js" type="text/javascript"></script>
+<script type="module" src="/js/recommended.js"></script>
 </body>
 </html>
